invoice_generation/requirements.txt
python-docx
pandas
docx2pdf
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
oauth2client
numpy